wip: decant pom wars commands
