+++ pombot/commands/pom_wars/__init__.py
@@ -1,0 +1,4 @@
+from pombot.commands.pom_wars.actions import *
+from pombot.commands.pom_wars.attack import *
+from pombot.commands.pom_wars.bribe import *
+from pombot.commands.pom_wars.defend import *
