+++ pombot/extensions/general.py
@@ -1,57 +1,45 @@
+import logging
 from functools import partial
 
-import discord
-from discord.ext.commands import Command, Context
-from discord.ext.commands.errors import MissingAnyRole, NoPrivateMessage
-from discord.ext.commands.bot import Bot
+from discord.ext.commands import Bot, Command
 
-import pombot
+from pombot import commands
+from pombot import listeners
 from pombot.config import Config
+from pombot.lib.tiny_tools import has_any_role
 
-
-def _has_any_role(ctx: Context, roles_needed=None) -> bool:
-    """A non-decorator reimplementation of discord.ext.commands.has_any_role,
-    but with dignity.
-    """
-    roles_needed = roles_needed or []
-
-    if not isinstance(ctx.channel, discord.abc.GuildChannel):
-        raise NoPrivateMessage()
-
-    get_user_roles = partial(discord.utils.get, ctx.author.roles)
-
-    if not any(get_user_roles(id=role_needed) is not None
-            if isinstance(role_needed, int)
-            else get_user_roles(name=role_needed) is not None
-                for role_needed in roles_needed):
-        raise MissingAnyRole(roles_needed)
-
-    return True
+_log = logging.getLogger(__name__)
 
 
 def setup(bot: Bot):
     """Load general commands and listeners."""
     admin = {
-        "checks": [partial(_has_any_role, roles_needed=Config.ADMIN_ROLES)]
+        "checks": [partial(has_any_role, roles_needed=Config.ADMIN_ROLES)]
     }
 
     for command in [
-        Command(pombot.commands.do_events,       name="events"),
-        Command(pombot.commands.do_howmany,      name="howmany"),
-        Command(pombot.commands.do_newleaf,      name="newleaf"),
-        Command(pombot.commands.do_pom,          name="pom"),
-        Command(pombot.commands.do_poms,         name="poms"),
-        Command(pombot.commands.do_reset,        name="reset"),
-        Command(pombot.commands.do_undo,         name="undo"),
+        Command(commands.do_events,       name="events"),
+        Command(commands.do_howmany,      name="howmany"),
+        Command(commands.do_newleaf,      name="newleaf"),
+        Command(commands.do_pom,          name="pom"),
+        Command(commands.do_poms,         name="poms"),
+        Command(commands.do_reset,        name="reset"),
+        Command(commands.do_undo,         name="undo"),
 
-        Command(pombot.commands.do_total,        name="total",        **admin),
-        Command(pombot.commands.do_create_event, name="create_event", **admin),
-        Command(pombot.commands.do_remove_event, name="remove_event", **admin),
+        Command(commands.do_total,        name="total",        **admin),
+        Command(commands.do_create_event, name="create_event", **admin),
+        Command(commands.do_remove_event, name="remove_event", **admin),
     ]:
         bot.add_command(command)
 
+    on_ready_handler = partial(listeners.handle_on_ready, log=_log)
+
     for listener in [
-        (partial(pombot.listeners.handle_on_ready, bot), "on_ready"),
-        (pombot.listeners.handle_on_command_error,       "on_command_error"),
+        # FIXME: - logs and prints don't show up
+        #           - passing in a 2nd arg does not work as expected
+        #        - the debugger can't step through the on-ready handlers.
+        #           - who the fuck wrote this garbage?
+        (partial(on_ready_handler, bot), "on_ready"),
+        (listeners.handle_on_command_error,       "on_command_error"),
     ]:
         bot.add_listener(*listener)
