+++ pombot/lib/pom_wars/scoreboard.py
@@ -1,4 +1,3 @@
-import logging
 from typing import List
 
 import discord.errors
@@ -7,15 +6,12 @@ from discord.ext.commands.bot import Bot
 
 from pombot.config import Pomwars, Reactions
 from pombot.lib.messages import send_embed_message
-from pombot.state import State
 from pombot.lib.pom_wars.team import Team
-
-_log = logging.getLogger(__name__)
 
 
 class Scoreboard:
     """A representation of the scoreboard in join channels."""
-    def __init__(self, bot, scoreboard_channels) -> None:
+    def __init__(self, bot: Bot, scoreboard_channels: List) -> None:
         self.bot = bot
         self.scoreboard_channels = scoreboard_channels
 
@@ -133,24 +129,3 @@ class Scoreboard:
                 restricted_channels.append(channel)
 
         return [full_channels, restricted_channels]
-
-
-async def setup_pomwar_scoreboard(bot: Bot):
-    """Find and remember the static scoreboard for all connected guilds."""
-    channels = []
-
-    for guild in bot.guilds:
-        for channel in guild.channels:
-            if channel.name == Pomwars.JOIN_CHANNEL_NAME:
-                channels.append(channel)
-
-    State.scoreboard = Scoreboard(bot, channels)
-    full_channels, restricted_channels = await State.scoreboard.update()
-
-    for channel in full_channels:
-        _log.error("Join channel '%s' on '%s' is not empty",
-            channel.name, channel.guild.name)
-
-    for channel in restricted_channels:
-        _log.error("Join channel '%s' on '%s' is not messagable (Missing Access)",
-            channel.name, channel.guild.name)
