+++ pombot/listeners/__init__.py
@@ -1,2 +1,3 @@
 from pombot.listeners.on_command_error import *
 from pombot.listeners.on_ready import *
+from pombot.listeners.pom_wars import *
