+++ pombot/listeners/pom_wars/__init__.py
@@ -1,0 +1,1 @@
+from pombot.listeners.pom_wars.on_ready import handle_on_ready
